nn_js_ws/fast-scnn/requirements.txt
absl-py==0.8.0
albumentations==0.3.2
astor==0.8.0
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
Cython==0.29.13
decorator==4.4.0
gast==0.2.2
gluoncv==0.5.0
graphviz==0.8.4
grpcio==1.23.0
h5py==2.9.0
idna==2.8
imageio==2.5.0
imgaug==0.2.6
joblib==0.13.2
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
mock==3.0.5
mxnet-mkl==1.5.0
networkx==2.3
numpy==1.17.1
opencv-python==4.1.1.26
opencv-python-headless==4.1.1.26
pandas==0.25.1
Pillow==6.1.0
protobuf==3.9.1
pycocotools==2.0.0
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyWavelets==1.0.3
PyYAML==5.1.2
requests==2.22.0
scikit-image==0.15.0
scipy==1.3.1
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
tqdm==4.35.0
urllib3==1.25.3
Werkzeug==0.15.5
